refactor: extract hook module interfaces into dedicated types.ts
