+++ libs/install/platforms/types.ts
@@ -1,5 +1,5 @@
 import type { InstallPlatform } from "../paths.js";
-import type { HookInput } from "../../hooks/hook-input.js";
+import type { HookInput } from "../../hooks/types.js";
 
 export interface PlatformInstallContext {
   repoRoot: string;
